refactor proof verification helper functions to a separate submodule
